lift wallet connect import logic to context
